Move `hash_items` into `specs`
We really don't like having an object_utils grab bag. This moves
`hash_items` to the root case file it's used in, thus allowing us
to once and for all remove `object_utils`
